darwin: bpf_program is uint32 not uint16
in MacOS man bpf page

struct bpf_program {
    u_int bf_len;
    struct bpf_insn *bf_insns;
};

otherwise you will get lots of invalid BPF command for ioctl

+++ pcap.go
@@ -27,7 +27,7 @@ type Packet struct {
 }
 
 type BpfProgram struct {
-	Len    uint16
+	Len    uint32
 	Filter *bpf.RawInstruction
 }
